mac_requirements.txt
Flask==2.3.2
psutil==5.9.5
pywebview==4.2.2

